Move `PluralForm` into a new module
Also replace the `panic` that could happen when an unknown plural
formula is used with returning `None`.
